Removed sidebar component
